Updates for monitoring the runtime metrics of underlying platform (Spark)
